Loading both Vim rails and projectionist plugins was causing issues with FZF `:Tags`.
I don't know why.

After some thought it is not best to have both plugins loaded, better
one or the other or none.

New scheme uses vim-plug delayed loading:

[*] load vim-rails only for Ruby and eRuby files

[*] load vim-projectionist only for Elixir, eElixir and Javascript/JSX
files

This has the side benefit of speeding up Vim/Neovim loading for projects
that are not Rails, or Elixir/Phoenix or React.