Disable Bash 5.1 colored paste
Bash 5.1 now enables bracketed-paste which is connected to coloured
pasting. This is ugly and distracting.

The only way to disable the highlighted paste is to disable
bracketed-paste. So be it.

Hopefully in future these two features (bracketed-paste and highlight
paste) can be separated.

See: https://is.gd/fsgOMe & https://is.gd/Uccmae