Ruby/Rails (Neo)vim speed improvements
vim-rails and vim-bundler plugins slow down the loading of Ruby files
(by a lot). Likewise don't load the Vim Ruby provider.

Use vim-projectionist to get back most (but not all) of vim-rails
functionality. The missing pieces are not worth the performance loss.

Loading of Ruby files is now about twice as fast as it used to be (300ms
vs 600ms).