Use neovim-remote for Neovim lazygit
lazygit will be launched from a floating window. This will result, by
default, in Neovim being launched within Neovim for edit and commit
operations. This is problematic since my 'Ctrl-q' mapping (to quit
Neovim) will sometimes cause the host Neovim to close rather than the
embedded Neovim.

To avoid this class of problems it is best not to launch Neovim inside a
floating terminal window; instead use neovim-remote to launch Neovim
(from lazygit) in a separate tab window. Once the tab is close control
will return to lazygit (in a floating window).

In practice this works well.