Fix flashing cursor in Neo-tree
'git_status_async = false' results in cursor flickering when editing code
with Neo-tree open. This option was set to enable automatic updates of
Git status symbols when staging code.

But the flicking cursor is annoying, so instead remove
'git_status_async' altogether and instead create an autocmd when
entering the Neo-tree do a Git status refresh.